+++ src/name-uniqueness-check.ts
@@ -1,38 +1,19 @@
 import { error, highlight } from "./chalk-formatting";
 import { ValidationErrorCodes } from "./error-codes";
-import { CustomType, InputParameters, OutputBranches, OutputData } from "./meta-function-type";
-
-type NamedElement = OutputData | CustomType | InputParameters
 
 /**
- * Checks if all provided Elements have a unique name
+ * Checks if all provided Elements have a unique property
  */
-export const nameUniquenessCheck = (namedElementList : NamedElement[], entityName : string) => {
+export const propertyUniquenessCheck =
+ <T extends object>(namedElementList : T[], propertyName : keyof T, entityName : string) => {
   const nameList = [];
 
   namedElementList.forEach((namedElement) => {
-    if (nameList.includes(namedElement.name)) {
+    if (nameList.includes(namedElement[propertyName])) {
       throw Error(error(ValidationErrorCodes.V27) +
-        `- "${entityName}" - ${highlight(`"${namedElement.name}"`)}`);
+        `- "${entityName}" - ${highlight(`"${namedElement[propertyName]}"`)}`);
     }
 
-    nameList.push(namedElement.name);
+    nameList.push(namedElement[propertyName]);
   })
 }
-
-/**
- * Checks if all provided outputBranches have a unique name
- */
-export const branchesNameUniquenessCheck = (namedElementList : OutputBranches[]) => {
-  const nameList = [];
-
-  namedElementList.forEach((namedElement) => {
-    if (nameList.includes(namedElement.branchName)) {
-      throw Error(error(ValidationErrorCodes.V27) +
-        `- "outputBranches" - ${highlight(`"${namedElement.branchName}"`)}`);
-    }
-
-    nameList.push(namedElement.branchName);
-  })
-}
-