+++ src/validate-string-configuration.ts
@@ -2,7 +2,8 @@ import { isMetaFunction } from "./is-meta-function";
 import { ValidationErrorCodes } from "./error-codes";
 import { MetaCustomTypesValidation } from "./custom-types-validation";
 import { error, success } from "./chalk-formatting";
-import { nameUniquenessCheck } from "./name-uniqueness-check";
+import { propertyUniquenessCheck } from "./name-uniqueness-check";
+import { validateOutputConfiguration } from "./validate-output-configuration";
 
 /** Validates the string content of a `meta-function.json` file */
 export const validateStringConfiguration = (configurationData : string) => {
@@ -15,9 +16,17 @@ export const validateStringConfiguration = (configurationData : string) => {
   }
 
   isMetaFunction(objectResult);
-  nameUniquenessCheck(objectResult.inputParameters, "inputParameters");
-  nameUniquenessCheck(objectResult.outputData, "outputData");
-  nameUniquenessCheck(objectResult.customTypes, "customTypes");
+  propertyUniquenessCheck(objectResult.inputParameters, "name", "inputParameters");
+  propertyUniquenessCheck(objectResult.outputData, "name", "outputData");
+  propertyUniquenessCheck(objectResult.customTypes, "name", "customTypes");
+  propertyUniquenessCheck(objectResult.outputBranches, "branchName", "outputBranches");
+
+  objectResult.customTypes.forEach((customType) => 
+    propertyUniquenessCheck(customType.properties, "name", "customType.properties"))
+
   new MetaCustomTypesValidation(objectResult).execute();
+
+  validateOutputConfiguration(objectResult.outputBranches, objectResult.outputData);
+
   console.log(success("File passed validation."));
 }